+++ src/main/java/com/netflix/zeno/fastblob/record/FastBlobDeserializationRecord.java
@@ -35,11 +35,11 @@ public class FastBlobDeserializationRecord implements NFDeserializationRecord {
     private final FastBlobSchema schema;
 
     private final ByteData byteData;
-    private final int fieldPointers[];
+    private final long fieldPointers[];
 
     public FastBlobDeserializationRecord(FastBlobSchema schema, ByteData byteData) {
         this.schema = schema;
-        this.fieldPointers = new int[schema.numFields()];
+        this.fieldPointers = new long[schema.numFields()];
         this.byteData = byteData;
     }
 
@@ -49,8 +49,8 @@ public class FastBlobDeserializationRecord implements NFDeserializationRecord {
      * @param objectBeginOffset
      * @return The length of the object's data, in bytes.
      */
-    public int position(int objectBeginOffset) {
-        int currentPosition = objectBeginOffset;
+    public int position(long objectBeginOffset) {
+        long currentPosition = objectBeginOffset;
 
         for(int i=0;i<fieldPointers.length;i++) {
             fieldPointers[i] = currentPosition;
@@ -60,7 +60,7 @@ public class FastBlobDeserializationRecord implements NFDeserializationRecord {
             currentPosition += fieldLength(currentPosition, type);
         }
 
-        return currentPosition - objectBeginOffset;
+        return (int)(currentPosition - objectBeginOffset);
     }
 
     /**
@@ -73,7 +73,7 @@ public class FastBlobDeserializationRecord implements NFDeserializationRecord {
     /**
      * get the offset into the byte data for the field represented by the String.
      */
-    public int getPosition(String fieldName) {
+    public long getPosition(String fieldName) {
         int fieldPosition = schema.getPosition(fieldName);
 
         if(fieldPosition == -1)
@@ -92,7 +92,7 @@ public class FastBlobDeserializationRecord implements NFDeserializationRecord {
         return fieldLength(fieldPointers[fieldPosition], fieldType);
     }
 
-    private int fieldLength(int currentPosition, FieldType type) {
+    private int fieldLength(long currentPosition, FieldType type) {
         if(type.startsWithVarIntEncodedLength()) {
             if(VarInt.readVNull(byteData, currentPosition)) {
                 return 1;
